requirements.txt
django==3.1.1
django-cors-headers==3.5
djangorestframework==3.12